Allow serial port to be optional
Useful for the microservices router instances, which don't need a
serial port anymore. Originally the serial port side of the router
was for connecting to the telemetry tx/rx processes, however we no
longer use those and everything is tcp/udp.

When we begin using this router for plain telemetry the serial port
connection will be for the actual UART pins on the air and ground.